% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_form_works.R
\name{check_form_works}
\alias{check_form_works}
\title{Check If The Form Is Sending Correctly}
\usage{
check_form_works(r)
}
\arguments{
\item{r}{response from HTTP POST request}
}
\description{
Check the HTTP status message is below 400, meaning that the transmission worked.
}
\examples{
\dontrun{
why_R_u_opensource("R is Cool!") \%>\% check_form_works()
# "All good."
}
}

