% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form.R
\name{get_form}
\alias{get_form}
\title{Scrape Google Form Content}
\usage{
get_form(form)
}
\arguments{
\item{form}{Can be either the form_url or form_id}
}
\description{
Get Google Form content by scraping page
}
\examples{
\dontrun{
url \%>\% get_form() -> scrape
}
}

