% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_form_entry_ids.R
\name{get_form_entry_ids}
\alias{get_form_entry_ids}
\title{Get Google Form Entry Ids From Scrape}
\usage{
get_form_entry_ids(scrape)
}
\arguments{
\item{scrape}{HTML Output from Scraping Google Form}
}
\description{
Get Google Form Entry Ids From Scrape
}
\examples{
\dontrun{
url \%>\% get_form() \%>\% get_form_entry_ids() -> entries
}
}

