% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_form_id.R
\name{get_form_id}
\alias{get_form_id}
\title{Extract Form Id From Form URL}
\usage{
get_form_id(url)
}
\arguments{
\item{url}{Google Form URL}
}
\description{
Extract Form Id From Form URL
}
\examples{
\dontrun{
url \%>\% get_form_id() -> id
}
}

