% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_form_questions.R
\name{get_form_questions}
\alias{get_form_questions}
\title{Extracts Google Form Question Text}
\usage{
get_form_questions(scrape)
}
\arguments{
\item{scrape}{HTML Output from Scraping Google Form}
}
\description{
Extracts Google Form Question Text From Scrape
}
\examples{
\dontrun{
url \%>\% get_form() \%>\% get_form_questions() -> questions
}
}

