% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form_str.R
\name{get_form_str}
\alias{get_form_str}
\title{Gets and Prints Form Questions and Entry.Ids}
\usage{
get_form_str(form)
}
\arguments{
\item{form}{The url to your form}
}
\description{
Gets and Prints Form Questions and Entry.Ids.
}
\examples{
\dontrun{
get_form_str("url_to_your_form")
}
}

