% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleformr-package.R
\docType{package}
\name{googleformr}
\alias{googleformr}
\alias{googleformr-package}
\alias{package-googleformr}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

