% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_answers.R
\name{post_answers}
\alias{post_answers}
\title{Posts Answers to a Google Form}
\usage{
post_answers(form, answers)
}
\arguments{
\item{form}{form id or url}

\item{answers}{vector of answers}
}
\description{
Posts Answers to a Google Form specified by url or id.
}
\examples{
\dontrun{
post_answers(form = "form_url",
             answers = c("Anything", "you", want))
}
}

