% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gformr.R
\name{gformr}
\alias{gformr}
\title{Creates Function Linking Form Entries To Post Content}
\usage{
gformr(form)
}
\arguments{
\item{form}{Can be either the form_url or form_id}
}
\description{
Creates Function Linking Form Entries To Post Content. This function allows you to embed
}
\examples{
\dontrun{
url \%>\% gformr() -> new_post_to_form
new_post_to_form(content_to_post_to_url)
}
}

