% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why_R_u_opensource.R
\name{why_R_u_opensource}
\alias{why_R_u_opensource}
\title{Sends Data back to my Google Form}
\usage{
why_R_u_opensource(post_content)
}
\arguments{
\item{post_content}{form Can be either the form_url or form_id}
}
\description{
Sends Data back to my Google Form About Being opensource. gformr creates why_R_u_opensource into a linked function with the form.
}
\examples{
\dontrun{
why_R_u_opensource(post_content = "the opensource community is helpful")
}
}

