% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/why_R_u_thankful.R
\name{why_R_u_thankful}
\alias{why_R_u_thankful}
\title{Sends Data back to my Google Form}
\usage{
why_R_u_thankful(post_content)
}
\arguments{
\item{form}{Can be either the form_url or form_id}
}
\description{
Sends Data back to my Google Form About Being Thankful. gformr creates why_R_u_thankful into a linked function with the form.
}
\examples{
\dontrun{
why_R_u_thankful("Aside from divine providence, the humbling, amazing transition I made from asst prof to data scientist was humanly possible bec of R and the opensource community. I love who I am now. I wouldn't go back.")
}
}

